Updated routes for MVC
