+++ package.json
@@ -24,6 +24,7 @@
     "body-parser": "^1.20.0",
     "connect-mongodb-session": "^3.1.1",
     "cors": "^2.8.5",
+    "dotenv": "^16.0.2",
     "express": "^4.18.1",
     "express-session": "^1.17.3",
     "mongodb": "^4.9.1",
